remove the component from component folder to move it to Main page folder
