+++ client/src/components/MainPage/TranslationsPage/DonateModal/TextTranslation/style.css
