+++ client/src/components/MainPage/TranslationsPage/DonateModal/VideoTranslation/style.css
