--- client/src/components/QuestionsPage/style.css
@@ -1,53 +1,0 @@
-.main {
-  width: 100%;
-  height: 100%;
-  display: grid;
-  grid-template-columns: .5fr 3fr;
-  grid-gap: 5px;
-
-}
-
-.main__content{
-  margin: 1rem;
-  align-items: center;
-  background-color: white;
-  padding: 2rem;
-  box-shadow: 0px 0px 23px -5px rgba(0, 0, 0, 0.75);
-
-}
-
-.main__content--text {
-  text-align: center;
-  color: #287fb8;
-  font-weight: bold;
-  font-size: 1.1rem;
-  padding-top: 10px;
-  cursor: pointer;
-}
-
-.main__content--text:hover {
-  color: #f59036;
-}
-
-.main__content--header{
-display: flex;
-justify-content: space-between;
-}
-
-.main__content--header--text {
-  color:#bc3636;
-}
-
-.main__content--header--search{
-  margin: 10px 20px;
-}
-
-@media (max-width: 1024px) {
-  .main {
-    display: grid;
-    grid-template-columns:  1fr;
-    grid-gap: 5px;
-  
-  }
-
-}
